test(Test): add uncovered test cases for TreeTable
